Move WebSocket implementation
